libs/mktdata/pyproject.toml
[tool.poetry]
name = "mktdata"
version = "1.0.0"
description = "MWZ mktdata utilities"
authors = ["Muwazana S.A. <exchanges@muwazana.com>"]

[tool.poetry.dependencies]
python = "^3.7"
aioredis = "^1.3.1"
python-json-logger = "^0.1.11"
google-cloud-storage = "^1.27.0"
pandas = "^1.0.3"
pyarrow = "^0.16.0"
gcsfs = "^0.6.1"

[tool.poetry.dev-dependencies]
black = { version = "*", allow-prereleases = true }
ipython = "^7.13.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
